requirements.txt
Django==2.0.7
django-betterforms==1.2
django-compressor==2.2
django-dynamic-fixture==2.0.0
django-test-plus==1.1.1
django-widget-tweaks==1.4.2
psycopg2-binary==2.7.4
